% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cipnumber_creation}
\alias{cipnumber_creation}
\title{Cipnumber Creation}
\usage{
cipnumber_creation(header, nrep)
}
\description{
Cipnumber Creation
}
\section{Functions}{
\itemize{
\item \code{cipnumber_creation}: to every cipnumber and repetition, it creates new cipnumber
}}

