% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_createlist.R
\name{create_ui}
\alias{create_ui}
\title{UI germoplasm and material list managment}
\usage{
create_ui(type = "tab", title = "Create List", name = "createList")
}
\arguments{
\item{type}{type of UI element, deault is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}
}
\description{
Returns user friendly ui
}
\author{
Omar Benites
}

