% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_createlist.R
\name{createlist_ui}
\alias{createlist_ui}
\title{UI Side for Creation Material List from Scratch}
\usage{
createlist_ui(type = "tab", title = "Family List", name = "createList")
}
\arguments{
\item{type}{type of UI element, deault is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}
}
\description{
Returns user friendly ui
}
\author{
Omar Benites
}

