% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_distributiondb.R
\name{distribution_ui}
\alias{distribution_ui}
\title{UI for Distribution DataBase}
\usage{
distribution_ui(type = "tab", title = "Distribution Database",
  name = "distributionDB")
}
\arguments{
\item{type}{type of UI element, deault is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}
}
\description{
Returns user friendly ui
}
\author{
Omar Benites
}
