% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbmlist_server.R
\name{fbmlist_server}
\alias{fbmlist_server}
\title{Server germoplasm and material list managment}
\usage{
fbmlist_server(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
Server germoplasm and material list managment
}
\author{
Omar Benites
}

