% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbmlist_ui.R
\name{fbmlist_ui}
\alias{fbmlist_ui}
\title{UI germoplasm and material list managment}
\usage{
fbmlist_ui(type = "tab", title = "Germoplasm Managment",
  name = "fbmlist_ui")
}
\arguments{
\item{type}{type of UI element, deault is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}
}
\description{
Returns user friendly ui
}
\author{
Omar Benites
}

