% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_generate.R
\name{generate_ui}
\alias{generate_ui}
\title{UI germoplasm and material ist managment Generation}
\usage{
generate_ui(type = "tab", title = "Generate List", name = "generateList")
}
\arguments{
\item{type}{type of UI element, deault is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}
}
\description{
Returns user friendly ui
}
\author{
Omar Benites
}

