% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_generate_agrofims.R
\name{generate_ui_agrofims}
\alias{generate_ui_agrofims}
\title{UI Side for generation of Material List in HIDAP AGROFIMS}
\usage{
generate_ui_agrofims(type = "tab", title = "Clone List",
  name = "generateList")
}
\arguments{
\item{type}{type of UI element, deault is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}
}
\description{
Returns user friendly ui
}
\author{
Omar Benites
}

