% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_managerlist.R
\name{managerlist_ui}
\alias{managerlist_ui}
\title{UI for open books in HiDAP
Returns user friendly ui}
\usage{
managerlist_ui(type = "tab", title = "Manage List", name = "manageList")
}
\arguments{
\item{type}{type of UI element, deault is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}
}
\description{
UI for open books in HiDAP
Returns user friendly ui
}
\author{
Omar Benites
}

