% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_parental.R
\name{parent_ui}
\alias{parent_ui}
\title{UI Side for parental of material list}
\usage{
parent_ui(type = "tab", title = "Parental List", name = "parentList")
}
\arguments{
\item{type}{type of UI element, deault is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}
}
\description{
Returns user friendly ui
}
\author{
Omar Benites
}
