% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fbmlist_data}
\alias{fbmlist_data}
\alias{mtl_files}
\title{Stablish conection to Material List Data Base}
\usage{
fbmlist_data()

mtl_files()
}
\description{
Stablish conection to Material List Data Base

List of the dbf files
}
\section{Functions}{
\itemize{
\item \code{fbmlist_data}: the table from M.List DB an write into the computer

\item \code{mtl_files}: the table from M.List DB an write into the computer
}}

