% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_createlist.R
\name{server_create}
\alias{server_create}
\title{Server germoplasm and material list managment}
\usage{
server_create(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
Server germoplasm and material list managment
}
\author{
Omar Benites
}

