% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_createlist.R
\name{server_createlist}
\alias{server_createlist}
\title{Server Side for Creation Material List from Scratch}
\usage{
server_createlist(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
Server Side for Creation Material List from Scratch
}
\author{
Omar Benites
}

