% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_distribution.R
\name{server_distribution}
\alias{server_distribution}
\title{Server side for distribution database}
\usage{
server_distribution(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
Server side for distribution database
}
\author{
Omar Benites
}

