% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_generate.R
\name{server_generate}
\alias{server_generate}
\title{Server Germoplasm and Material List Managment}
\usage{
server_generate(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
Server Germoplasm and Material List Managment
}
\author{
Omar Benites
}

