% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_generate_agrofims.R
\name{server_generate_agrofims}
\alias{server_generate_agrofims}
\title{Server Side for generation of Material List in HIDAP-AGROFIMS}
\usage{
server_generate_agrofims(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
Server Side for generation of Material List in HIDAP-AGROFIMS
}
\author{
Omar Benites
}

