% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_managerlist.R
\name{server_managerlist}
\alias{server_managerlist}
\title{Server openbooks}
\usage{
server_managerlist(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
Server openbooks
}
\author{
Omar Benites
}

