% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_managerlist_agrofims.R
\name{server_managerlist_agrofims}
\alias{server_managerlist_agrofims}
\title{Server openbooks in HIDAP-AGROFIMS}
\usage{
server_managerlist_agrofims(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
Server openbooks in HIDAP-AGROFIMS
}
\author{
Omar Benites
}

