% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_parental.R
\name{server_parentlist}
\alias{server_parentlist}
\title{Server Side for parental of material list}
\usage{
server_parentlist(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
Server Side for parental of material list
}
\author{
Omar Benites
}

