% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_utils_sql.R
\name{cipnumber_creation}
\alias{cipnumber_creation}
\title{Cipnumber Creation}
\usage{
cipnumber_creation(header, nrep)
}
\arguments{
\item{header}{header}

\item{nrep}{number of repetitions}
}
\description{
Cipnumber Creation
}
\section{Functions}{
\itemize{
\item \code{cipnumber_creation}: to every cipnumber and repetition, it creates new cipnumber
}}

