% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_utils.R
\name{accessname_code}
\alias{accessname_code}
\title{Accession Name Creation}
\usage{
accessname_code(accession_name, accession_number, nrep)
}
\arguments{
\item{accession_name}{the accession's name}

\item{nrep}{number of repetitions}

\item{accesssion_number}{the accession's number}
}
\description{
Accession Name Creation
}
\section{Functions}{
\itemize{
\item \code{accessname_code}: of accession names according to the number of repetitions
}}

