% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{accessname_code}
\alias{accessname_code}
\title{Accession Name Creation}
\usage{
accessname_code(accession_name, nrep)
}
\description{
Accession Name Creation
}
\section{Functions}{
\itemize{
\item \code{accessname_code}: of accession names according to the number of repetitions
}}

